# Regenerate all protobuf messages (#6082)