… src/contrib/cluster/Akka.Cluster.Sharding/Serialization/Proto/ClusterShardingMessages.g.cs
… src/contrib/cluster/Akka.Cluster.Tools/Client/Serialization/Proto/ClusterClientMessages.g.cs
… src/contrib/cluster/Akka.Cluster.Tools/PublishSubscribe/Serialization/Proto/DistributedPubSubMessages.g.cs
… src/contrib/cluster/Akka.DistributedData/Serialization/Proto/ReplicatedDataMessages.g.cs
… src/contrib/cluster/Akka.DistributedData/Serialization/Proto/ReplicatorMessages.g.cs
… src/core/Akka.Cluster/Serialization/Proto/ClusterMessages.g.cs
… src/core/Akka.Persistence/Serialization/Proto/Persistence.g.cs
… src/core/Akka.Remote.TestKit/Proto/TestConductorProtocol.g.cs
… src/core/Akka.Remote/Serialization/Proto/ContainerFormats.g.cs
… src/core/Akka.Remote/Serialization/Proto/SystemMessageFormats.g.cs
… src/core/Akka.Remote/Serialization/Proto/WireFormats.g.cs
… src/core/Akka.Streams/Serialization/Proto/StreamRefMessages.g.cs
… tools/AzCopy/Newtonsoft.Json.dll
… tools/AzCopy/System.Spatial.dll
… tools/AzCopy/Third Party Notice (Microsoft Azure Storage Tools).rtf