A sort of "backport" of #6082
Regenerate the protobuf codes in `1.4` branch

# Benchmark Result
## Before
``` ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1415 (2004/May2020Update/20H1)
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=6.0.200
  [Host]     : .NET 6.0.2 (6.0.222.6406), X64 RyuJIT
  DefaultJob : .NET 6.0.2 (6.0.222.6406), X64 RyuJIT
```
|                 Method |       Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
|----------------------- |-----------:|---------:|---------:|-------:|-------:|----------:|
|        WritePayloadPdu | 1,669.6 ns | 21.10 ns | 19.74 ns | 0.2156 |      - |   1,808 B |
|       DecodePayloadPdu | 2,039.7 ns | 12.52 ns | 11.71 ns | 0.2156 | 0.0031 |   1,816 B |
|          DecodePduOnly |   131.3 ns |  1.32 ns |  1.11 ns | 0.0563 | 0.0002 |     472 B |
|      DecodeMessageOnly | 1,665.0 ns | 15.03 ns | 14.05 ns | 0.1406 |      - |   1,184 B |
| DeserializePayloadOnly |   151.2 ns |  1.88 ns |  1.76 ns | 0.0199 |      - |     168 B |
## After
``` ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1415 (2004/May2020Update/20H1)
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=6.0.200
  [Host]     : .NET 6.0.2 (6.0.222.6406), X64 RyuJIT
  DefaultJob : .NET 6.0.2 (6.0.222.6406), X64 RyuJIT
```
|                 Method |       Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Allocated |
|----------------------- |-----------:|---------:|---------:|-------:|-------:|----------:|
|        WritePayloadPdu | 1,623.4 ns | 19.95 ns | 18.66 ns | 0.2219 | 0.0031 |   1,880 B |
|       DecodePayloadPdu | 1,738.6 ns | 22.79 ns | 21.31 ns | 0.2250 |      - |   1,888 B |
|          DecodePduOnly |   175.1 ns |  2.31 ns |  1.93 ns | 0.0572 |      - |     480 B |
|      DecodeMessageOnly | 1,296.8 ns | 11.89 ns | 10.54 ns | 0.1469 | 0.0016 |   1,232 B |
| DeserializePayloadOnly |   143.6 ns |  1.59 ns |  1.33 ns | 0.0199 | 0.0002 |     168 B |